Fix for get_basf2_git_hash to work with new basf2 tools
With the new BelleII tools, it seems that `$BELLE2_RELEASE` is only set if actually a release is used and is left unset for a development setup, where it head been set to `head` in the past. As a result, I got `None` as a git hash when running this function.
I suppose it's no bug in the tools, so one could adapt this function here.