Download dataset on success in get_job_status method
When sending jobs to the grid with gbasf2, the their status can be checked with
CLI tools, but once they succeed, they stay on the grid until the user
explicitly downloads them with a separate command. When using a Gbasf2-wrapping
batch process, I want it to download the datasets automatically once they are
finished. I could override the run method and make my own job status query loop,
but that defeats the idea of letting luigi do the job status checking.

The simplest solution that I found and added in this PR is to add the download
in the `get_job_status` method when the job status is successful. However, I
don't really like this, because it goes against the expectation that
`get_job_status` does nothing but checking the job status, this is a very strong
side effect.