Create a jinja2 template for gbasf2 steering file wrapper
For sending jobs to the grid with gbasf2, I serialize the path and create a
steering file to send to the grid which does nothing but read in the path and
process it. Because I want to influence some variable values in that file I had
it previously as a string which I then formatted and wrote to a file, but I
think it's better to have that as a template which people can view and edit
seperately from the batch process.