Require users to define project name as b2luigi param or attribute
With a setting, it would be the same for all instances of a task with different
parameters. Ideally, we want the project to be unique for different task
instances. As long as we don't find a way to encode the parameters in the
project name, it will be the user's task to check for that.