Hash the entire task id for the unique project name suffix
Before I just took the last part of the task ID string, which is a hash function
of the significant luigi parameters of the task. However, we want to ensure
different project names when we have different tasks with the same parameters,
so one should hash the entire task id.