Raise early error with useful message gbasf2 submission fails
Often, the subprocess command to submit a gbasf2 job can fail without an error
code thrown by gbasf2 (e.g. b/c of online problems on the grid),
so the `check=True` doesn't raise a CalledProcessError.
The `get_job_status` method always checks first if the project exists on the
grid and it will then see that it doesn't. In the past it just returned
JobStatus.idle, which results in the b2luigi error:

    ValueError("get_job_status() returned an unknown job state!")

When the user sees that, its hard to understand for him that the cause of this
error was a failed job gbasf2. So I introduced a more informative error message.