Move gbasf2 helper functions from within class to outside
I am thinking of defining a luigi Targed for project outputs on grid storage and
for that class I would need many of the same helper functions that I used in the
gbasf2 batch process. These functions, that basically wrap gbasf2 commands,
don't depend much on the task they are in, so I can redefine them as free functions.