Provide luigi target for gbasf2 project output datasets
If the user disables dataset download from the grid, he can't use the downloaded
dataset as a file system target and so has to check if the output exists on the
grid and if the project that creates that output is finished. This is
accomplished via a new target class.