Add run_with_gbasf2 function and rm gbasf2_env from class properties
`run_with_gbasf2()` is a free function that can be re-used in other code.
It calls the `get_gbasf2_env()` by itself to obtain the environment. Therefore,
I don't need that as a property anymore in my class. To avoid always calling the
same subprocess for sourcing the environment, I now cache the
`get_gbasf2_env()` function (again).