Change output for luigi tasks: Directories with names of contained root files
The dataset download from the grid now expects that the user provides the names
of the output root files via

        yield self.add_to_output("<name>.root")

It the batch process than iterates over the output and downloads all output files from
the project that match `<name>_*.root` into a directory named `<name>.root`.

The issue of that code as of now is, that in the create path the user currently
can't use `self.get_output_file_name` for defining his outputs, as only flat
output file names are supported in the output dataset.

Here's how the output of my example task that I use for testing currently looks like:

└── gbasf2_input_dataset=hashed_f2002a56f1ac44e1bfb6d3c0750cd955
    └── mbc_lower_cut=5.15
        ├── B_ntuple.root
        │   └── B_ntuple_0.root
        ├── D_ntuple.root
        │   └── D_ntuple_0.root
        └── MyAnalysisTask
            ├── serialized_basf2_path.pkl
            └── steering_file_wrapper.py