Use treeFit
`modularAnalysis.fitVertex` has been moved to `vertex.fitVertex` in release 4,
but that is not recommended to be used anymore. Instead, it is recommended to
use the new vertex fitter which is called `vertex.vertexTree` in release 4 and
then has been rename to `vertex.treeFit` in the light releases following release
4 (i.e. light-2002-janus). It also performs vertex fits for all daughter decays.

The docstring of `vertex.fitVertex` says:

    Direct use of `fitVertex` is not recommended unless you know what you are doing.
    If you're unsure, you probably want to use `treeFit` or `kFit`.

So I used treeFit with a try/except clause to enable both current versions,
depending on whether release 4 or light release of it is used.