Bugfix: Fix logic in setting gbasf2_additional_params (#43)
The intended way is that the value of the setting should be added to the gbasf2
submit command string only if the setting deviates from its default value which
is False. I had this pattern for multiple settings and here I had introduced a
copypaste error where whether this string is added to the string depends on
whether basf2opt is explicitly set.

Thanks for Max Welsch for finding this.

Co-authored-by: Michael Eliachevitch <meliache@uni-bonn.de>