Replace asserts with exceptions and improve some of the value checks
Further, I think despite stating that, the prior to this commit, there was a bug
when gbasf2_additional_files was e.g. a tuple of strings or another iterable of
strings, because I was trying to append that to a list. So I changed the code to
allow any iterables of strings and convert it to a string before appending