Allow dashes and underscores in gbasf2 project names (#45)
* Allow "-" or "_" in gbasf2 project names, but give a warning

* Replace asserts with exceptions and improve some of the value checks

Further, I think despite stating that, the prior to this commit, there was a bug
when gbasf2_additional_files was e.g. a tuple of strings or another iterable of
strings, because I was trying to append that to a list. So I changed the code to
allow any iterables of strings and convert it to a string before appending

* Fix typo/grammar in error message

* Fix regex logic for checking valid gbasf2 project name

Also mead it better self-documenting

* Improve error messages and refactor code to make logic more clear

* Remove warning if _ or - are used in gbasf2 project names

Reason is that Sviatoslav Bilokin said:
> From my personal experience, I never had any problems with the '-' symbol in project names.
> Also I asked responsible for user analysis in Computing Group, and he told me that '-' and '_' will be supported in the future, so there is no need to worry. And, probably, there is no need for warnings too

Co-authored-by: Michael Eliachevitch <meliache@uni-bonn.de>