New gbasf2 script to check for dirac proxy instead of string parsing
Before, we ran gb2_proxy_info and parsed the output of that command to see if
there is an alive proxy with a positive time left. But since that can break, now
use our own script that interacts with BelleDirac and has a returncode that
depends on whether the proxy is alive and with that it is easier to interact.