Run setup_dirac_proxy as a side-effect of run_with_gbasf2
When I looked at the code I saw that it is only run as a side effect of
``get_dirac_user``, but actually, we should ensure that the proxy is alive for
most gbasf2 commands.

The only exception is the gbasf2 subprocess to set up the dirac proxy, there I
had to be careful to prevent infinite loops.