Use Dirac API for Dirac proxy checking and checking for gbasf2 project existance (#51)
* New gbasf2 script to check for dirac proxy instead of string parsing

Before, we ran gb2_proxy_info and parsed the output of that command to see if
there is an alive proxy with a positive time left. But since that can break, now
use our own script that interacts with BelleDirac and has a returncode that
depends on whether the proxy is alive and with that it is easier to interact.

* Run setup_dirac_proxy as a side-effect of run_with_gbasf2

When I looked at the code I saw that it is only run as a side effect of
``get_dirac_user``, but actually, we should ensure that the proxy is alive for
most gbasf2 commands.

The only exception is the gbasf2 subprocess to set up the dirac proxy, there I
had to be careful to prevent infinite loops.

* Update documentation to code changes

* Make proxy initialized check script executable and fix subprocess call

* Make job status getter script executable

* Use up-to-date existing grid dataset for gbasf2 batch example

* Allow empty gb2 status strings as sign that grid job does not exist

Seems that's a new behavior in the new gbasf2 versions. Really should go away
from string parsing.

* Remove hint output that does not work anymore

* Remove unused import

* Use existing DIRAC API job status dict function to see if project exists

Before we did string parsing, but that is unstable

Co-authored-by: Michael Eliachevitch <meliache@uni-bonn.de>