Revert "Store htcondor job IDs in cache to avoid resubmitting tasks"
This reverts commit 3d7bfa121daa5a26b523d4decb6715dc917d0367.

This is embarassing, but I accidentally pushed that commit to the main branch.
It intended to push it on the feature branch

        feature/htcondor-check-for-running-jobs-before-submitting

These features that I pushed are still experimental and seemed to work for me,
but might break in some cases and it was not intended for publication yet. I
think for this I'd do a PR at some point.

Sorry for miss-using admin rights. I'll check my settings how I can prevent this
from happening in the future.