Feature/remove useless super delegation (#54)
* Remove useless super delegation

Again something that pylint gave me. This might be useful when wanting to change
the cache size explicitly. But if that's not used, maybe we can just remove
this.

On another note, in python3 it would have been also possible to write

    super().__init__()

without arguments.

* Give explicit check argument to subprocess.run & whitespace cleanup

Basically just implemented some more pylint recommendations.

Pylint recommends writing the the `check=False` explicitly into
`subprocess.run`, because the Zen of python tells us to be explicit when
silencing errors. Not really sure that's necessary here

Co-authored-by: Michael Eliachevitch <meliache@uni-bonn.de>