Fix broken code formatting in example
The code formatting for the example in the docs for
``get_input_file_names_from_dict()`` was broken, so I fixed that, by adding an
additional code-block environment.

Also I changed one other example for consistency, so that the text "Example" is
now the key/label of an RST definition list.