Fix validation of downloaded grid dataset for duplicate LFNs
For checking if a grid dataset has been completely downloaded, I compare the
output of `gb2_ds_list` (lists the remote dataset) with the list of downloaded
files. However, `gb2_ds_list` also lists duplicate outputs from jobs that had
been rescheduled. The `gb2_ds_get` download script filters those out, but with
this, I can't simply use the comparison anymore.

Thus, via this PR I manually filter out the duplicated remote LFN's and only
show those with the highest number. For the logic, I adapted some of the code
of gbasf2/BelleDirac, in particular the `findDuplicatedJobID()` function:
https://stash.desy.de/projects/B2DC/repos/belledirac/browse/gbasf2/lib/ds/util.py#214

Sadly, it also just uses string parsing. I was a little worried that it might
brake for old datasets produced with old gbasf2 versions, but I try to recognize
if the job LFN is the new version and if it's not I don't remove anything.

There is a gbasf2 issue to do this filtering automatically, see
https://agira.desy.de/browse/BIIDCD-1240, so after that's merged this code might
be simplified.

My wish is that it will be possible to just rely on the exit code of
`gb2_ds_get` to know if a dataset has been successfully and completely
downloaded. This whish has been also formulated into an issue: https://agira.desy.de/browse/BIIDCD-1238