Fix b2luigi __version__ being shadowed by luigi
The import statement

    from luigi import *

also imports the `__version__` from luigi and this seems to shadow the b2luigi version
defined at the top of the file.

I found this when I checked `b2luigi.__version__` on a remote host to check the
installed version and saw 3.0.2 as a result, which is the luigi version. I
wonder why I didn't see this earlier...