Never print all downloaded, but missing/superfluous grid files
Basically I removed the `if verbose` and just always print missing files if
there are some missing.

I realized that I don't really need the verbose parameter and with the printing
of difference implemented by Arthur, it's also not necessary to print the list
of all files. And anyway, if I can shorten code without any loss, it's better to
do so.

I don't think using many keyword parameters in private code is a good
design pattern anyway, it's not a flag meant for users.