Allow for gbasf2 projects with multiple output sub<xy> directories
Before that, the assumption was that all outputs are saved in a final sub00
directory. But from gbasf2 release v5r1p3 to be released on 2021-09-16, jobs
with many outputs (> 1000) can have outputs in additional subxy directories.

This means that when downloading, we will also get multiple sub<xy> directories in
the download. However, currently, the b2luigi user expects that all downloaded
file names will be in a common directory, so after a completed download, we move
the contents of all sub* directories into the final output directory.

We can then still cross-check that the download was complete by replace sub<xy>
with a wildcard when doing the remote gb2_ds_list command

See https://github.com/nils-braun/b2luigi/issues/80