Fix gbasf2 batch example for new basf2 releases: import ROOT
If ROOT is not explicitly imported, the example fails for the light-2106-rhea
release, at least for me. Because the stdCharged packages try to import
something from the ROOT namespace, but python complains that such a package does
not exist. Explicitly importing ROOT fixes this. Not sure if that's also a
problem for future basf2 releases or a bug in basf2.