Safer code for generating name of old monitoring downloads file
Doing "replace" on the file path for the current monitoring file might be
problematic, if the strings "failed_files" appears somewhere in the
filepath (e.g. a directory). In this case it is unlikely, but still, I prefer
using os.path.splitext to get the stem (part before the extension) of a filepath
and just append a suffix to that and re-append the extension.