Use with...open to free resources after reading file
In the past I myself used `open(...).readlines()`, but I found it is considered
better style to use `with` to explicitly free resources after opening the file.
In the first case the garbage collector would to it at some point.

Found this due to a pylint warning, see: https://stackoverflow.com/questions/67419064/r1732-consider-using-with-for-resource-allocating-operations-consider-using