Don't subpress CalledProcessError in get_proxy_info and add tests
When the proxy is not initalized, the proxy info script raises an error. Better
not subpress that, because in some cases we want the user to notice and error if
get_proxy_info is called before an initialization is happened. The except we can
do where we choose to ignore the error.