Update pre-commit flake8 version, fixing CI failure
Update `pre-commit` hooks. Most notably for the developers, update the `flake8`
syntax and style-checker to version 5.0.4, which might change slightly what
style is accepted. This should also fix [an
issue](https://github.com/python/importlib_metadata/issues/406) with the old
flake8 version not being compatible with the latest version of `importlib_meta`,
which the pre-commit flake8 hook in the github actions to fail. In the process
also migrated the pre-commit config format to the new layout.