Change implementation of task iterator to avoid adding parameter
Previously, the `already_seen_tasks` parameter was added so that in the
recursion it can be used to keep track of the state of already yielded tasks.
However, we can create such a cache by instatiating a new variable internally in
the function and then create a new internal function which accesses this state.