Bump to v0.9.0
The change to the `task_iterator` to give a unique list of tasks instead of
repeating tasks when nodes share children affects all remote batch submissions.
It should not break submission, just improve submission performance. But is a
change in the behavior of a public-facing function and therefore changing the
minor digit of the version is warranted.