Raise error if parameter value is not interpretable as basename
When the user defines a parameter that contains `/` (or `\` on windows), when
building the output path it will be interpreted as sub-directories. In the past
I had added a warning if the path contains `/`. But for Windows-compatibility
one should check for `os.path.sep` instead. Recently I got the idea that we can
check whether the serialized parameter value if identical to the
`os.bath.basename` of itself. This should exclude all cases where it's
interpreted as multiple directories.

Also I don't know why in the past I had just printed a warning. I think we can
be more aggressive and just raise a proper error. I think back then I was still
just a young contributor and worried to change too much. Still it

I'm very happy if I can find somebody to test and review my code, which was
difficult when I was maintaining b2uigi 😅. This is not really tested yet so
would be nice if somebody could to thit.